Add database in identity repository

+++ src/modules/identity/services/delete-user.service.ts
@@ -5,9 +5,13 @@ import { UserRepository } from "../repositories/user.repository";
 
 @Injectable()
 export class DeleteUserService {
-  constructor(private readonly userRepository: UserRepository) {}
+  constructor(private readonly userRepository: UserRepository) { }
 
   async execute(user_id: string): Promise<void> {
-        await this.userRepository.delete(user_id);
+    const userExists = await this.userRepository.findById(user_id);
+    if (!userExists) {
+      throw new Error("User not found");
+    }
+    await this.userRepository.delete(user_id);
   }
 }
